+++ rasa/core/domain.py
@@ -300,20 +300,20 @@ class Domain(object):
 
         return int(text_hash, 16)
 
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def user_actions_and_forms(self):
         """Returns combination of user actions and forms"""
 
         return self.user_actions + self.form_names
 
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def num_actions(self):
         """Returns the number of available actions."""
 
         # noinspection PyTypeChecker
         return len(self.action_names)
 
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def num_states(self):
         """Number of used input states for the action prediction."""
 
@@ -388,7 +388,7 @@ class Domain(object):
             return None
 
     # noinspection PyTypeChecker
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def slot_states(self) -> List[Text]:
         """Returns all available slot state strings."""
 
@@ -399,28 +399,28 @@ class Domain(object):
         ]
 
     # noinspection PyTypeChecker
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def prev_action_states(self) -> List[Text]:
         """Returns all available previous action state strings."""
 
         return [PREV_PREFIX + a for a in self.action_names]
 
     # noinspection PyTypeChecker
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def intent_states(self) -> List[Text]:
         """Returns all available previous action state strings."""
 
         return ["intent_{0}".format(i) for i in self.intents]
 
     # noinspection PyTypeChecker
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def entity_states(self) -> List[Text]:
         """Returns all available previous action state strings."""
 
         return ["entity_{0}".format(e) for e in self.entities]
 
     # noinspection PyTypeChecker
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def form_states(self) -> List[Text]:
         return ["active_form_{0}".format(f) for f in self.form_names]
 
@@ -429,12 +429,12 @@ class Domain(object):
 
         return self.input_state_map.get(state_name)
 
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def input_state_map(self) -> Dict[Text, int]:
         """Provides a mapping from state names to indices."""
         return {f: i for i, f in enumerate(self.input_states)}
 
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def input_states(self) -> List[Text]:
         """Returns all available states."""
 
@@ -687,7 +687,7 @@ class Domain(object):
         """Return the configuration for an intent."""
         return self.intent_properties.get(intent_name, {})
 
-    @common_utils.lazyproperty
+    @common_utils.lazy_property
     def intents(self):
         return sorted(self.intent_properties.keys())
