+++ tests/core/test_broker.py
@@ -1,7 +1,10 @@
 import json
 
-from rasa.core import broker
-from rasa.core.broker import FileProducer, PikaProducer, KafkaProducer, SQLProducer
+import rasa.core.brokers.utils as broker_utils
+from rasa.core.brokers.file_producer import FileProducer
+from rasa.core.brokers.kafka import KafkaProducer
+from rasa.core.brokers.pika import PikaProducer
+from rasa.core.brokers.sql import SQLProducer
 from rasa.core.events import Event, Restarted, SlotSet, UserUttered
 from rasa.utils.endpoints import EndpointConfig, read_endpoint_config
 from tests.core.conftest import DEFAULT_ENDPOINTS_FILE
@@ -17,7 +20,7 @@ def test_pika_broker_from_config():
     cfg = read_endpoint_config(
         "data/test_endpoints/event_brokers/pika_endpoint.yml", "event_broker"
     )
-    actual = broker.from_endpoint_config(cfg)
+    actual = broker_utils.from_endpoint_config(cfg)
 
     assert isinstance(actual, PikaProducer)
     assert actual.host == "localhost"
@@ -28,7 +31,7 @@ def test_pika_broker_from_config():
 def test_no_broker_in_config():
     cfg = read_endpoint_config(DEFAULT_ENDPOINTS_FILE, "event_broker")
 
-    actual = broker.from_endpoint_config(cfg)
+    actual = broker_utils.from_endpoint_config(cfg)
 
     assert actual is None
 
@@ -37,7 +40,7 @@ def test_sql_broker_from_config():
     cfg = read_endpoint_config(
         "data/test_endpoints/event_brokers/sql_endpoint.yml", "event_broker"
     )
-    actual = broker.from_endpoint_config(cfg)
+    actual = broker_utils.from_endpoint_config(cfg)
 
     assert isinstance(actual, SQLProducer)
     assert actual.engine.name == "sqlite"
@@ -47,7 +50,7 @@ def test_sql_broker_logs_to_sql_db():
     cfg = read_endpoint_config(
         "data/test_endpoints/event_brokers/sql_endpoint.yml", "event_broker"
     )
-    actual = broker.from_endpoint_config(cfg)
+    actual = broker_utils.from_endpoint_config(cfg)
 
     for e in TEST_EVENTS:
         actual.publish(e.as_dict())
@@ -64,7 +67,7 @@ def test_file_broker_from_config():
     cfg = read_endpoint_config(
         "data/test_endpoints/event_brokers/file_endpoint.yml", "event_broker"
     )
-    actual = broker.from_endpoint_config(cfg)
+    actual = broker_utils.from_endpoint_config(cfg)
 
     assert isinstance(actual, FileProducer)
     assert actual.path == "rasa_event.log"
@@ -73,7 +76,7 @@ def test_file_broker_from_config():
 def test_file_broker_logs_to_file(tmpdir):
     fname = tmpdir.join("events.log").strpath
 
-    actual = broker.from_endpoint_config(
+    actual = broker_utils.from_endpoint_config(
         EndpointConfig(**{"type": "file", "path": fname})
     )
 
@@ -92,7 +95,7 @@ def test_file_broker_logs_to_file(tmpdir):
 def test_file_broker_properly_logs_newlines(tmpdir):
     fname = tmpdir.join("events.log").strpath
 
-    actual = broker.from_endpoint_config(
+    actual = broker_utils.from_endpoint_config(
         EndpointConfig(**{"type": "file", "path": fname})
     )
 
@@ -110,13 +113,13 @@ def test_file_broker_properly_logs_newlines(tmpdir):
 
 
 def test_load_custom_broker_name():
-    config = EndpointConfig(**{"type": "rasa.core.broker.FileProducer"})
-    assert broker.from_endpoint_config(config)
+    config = EndpointConfig(**{"type": "rasa.core.brokers.file.FileProducer"})
+    assert broker_utils.from_endpoint_config(config)
 
 
 def test_load_non_existent_custom_broker_name():
-    config = EndpointConfig(**{"type": "rasa.core.broker.MyProducer"})
-    assert broker.from_endpoint_config(config) is None
+    config = EndpointConfig(**{"type": "rasa.core.brokers.my.MyProducer"})
+    assert broker_utils.from_endpoint_config(config) is None
 
 
 def test_kafka_broker_from_config():
